backend/requirements/development.txt
-r base.txt

# 개발 도구
django-debug-toolbar>=4.2.0
django-extensions>=3.2.0

# 테스팅
pytest>=7.4.0
pytest-django>=4.5.0
factory-boy>=3.3.0
coverage>=7.3.0

# 코드 품질
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# 개발 서버
watchdog>=3.0.0
